+++ packages/content/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @bluecadet/launchpad-content
 
+## 1.7.0
+
+### Minor Changes
+
+- [`f3c7916`](https://github.com/bluecadet/launchpad/commit/f3c79169001dd157c7e3bce24da41409a3906d53) Thanks [@benjaminbojko](https://github.com/benjaminbojko)! - Simplified json url parsing for better efficiency with large files
+
+## 1.6.0
+
+### Minor Changes
+
+- [#62](https://github.com/bluecadet/launchpad/pull/62) [`a4132da`](https://github.com/bluecadet/launchpad/commit/a4132da0187f669ad95251e2f3903229e87d6123) Thanks [@benjaminbojko](https://github.com/benjaminbojko)! - Added caching and blur to image transforms
+
+## 1.5.0
+
+### Minor Changes
+
+- [#61](https://github.com/bluecadet/launchpad/pull/61) [`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41) Thanks [@benjaminbojko](https://github.com/benjaminbojko)! - Refactored media downloads to support alt local paths. Adds support for Sanity image hotspots.
+
+### Patch Changes
+
+- Updated dependencies [[`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41)]:
+  - @bluecadet/launchpad-utils@1.2.1
+
+## 1.5.0-next.0
+
+### Minor Changes
+
+- [#61](https://github.com/bluecadet/launchpad/pull/61) [`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41) Thanks [@benjaminbojko](https://github.com/benjaminbojko)! - Refactored media downloads to support alt local paths. Adds support for Sanity image hotspots.
+
+### Patch Changes
+
+- Updated dependencies [[`f9b2140`](https://github.com/bluecadet/launchpad/commit/f9b21407af6d4f874473eed860e7a925475b7e41)]:
+  - @bluecadet/launchpad-utils@1.2.1-next.0
+
+## 1.4.0
+
+### Minor Changes
+
+- [#47](https://github.com/bluecadet/launchpad/pull/47) [`7f93181`](https://github.com/bluecadet/launchpad/commit/7f9318171c7d44ef812243454608d75810895d14) Thanks [@pingevt](https://github.com/pingevt)! - Updating Sanity source
+
+  Changes config for custom Queries
+  Adds an option for saving data in multiple files or in one file
+
+### Patch Changes
+
+- [#47](https://github.com/bluecadet/launchpad/pull/47) [`c9a66e1`](https://github.com/bluecadet/launchpad/commit/c9a66e1416d49c1447d010ab08b3de9c45b4e0a0) Thanks [@pingevt](https://github.com/pingevt)! - Add in better checking to process Sanity Text Blocks
+
 ## 1.3.0
 
 ### Minor Changes